update paceval api to handle url encoded string 2
