+++ examples_sources/NodeJS_examples/k8s/pacevalAPIService/main.go
@@ -20,7 +20,8 @@ func main() {
 	log.Info().Msg("starting service...")
 
 	singleProxyHandler := http2.NewSingleHostProxyHandler(manager)
-	multiRequestsHandler := http2.NewMultiHostProxyHandler(manager)
+	multiRequestHandler := http2.NewMultiHostRequestHandler(manager)
+	multiRequestsExtHandler := http2.NewMultiHostRequestExtHandler(manager)
 	demoHandler := http2.NewDemoHandler(manager)
 
 	http.Handle("/Demo/", demoHandler)
@@ -30,7 +31,8 @@ func main() {
 	http.Handle("/GetComputationResultExt/", singleProxyHandler)
 	http.Handle("/GetComputationInformationXML/", singleProxyHandler)
 	http.Handle("/GetErrorInformation/", singleProxyHandler)
-	http.Handle("/GetMultipleComputationsResults/", multiRequestsHandler)
+	http.Handle("/GetMultipleComputationsResults/", multiRequestHandler)
+	http.Handle("/GetMultipleComputationsResultsExt/", multiRequestsExtHandler)
 	http.HandleFunc("/healthz", func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Cache-Control", "no-cache, no-store, no-transform, must-revalidate, private, max-age=0")
 		w.WriteHeader(http.StatusNoContent)
